requirements.txt
streamlit
openai
langchain
langchain-openai
langchain-community
pdfplumber
pandas
openpyxl
python-docx
tiktoken
faiss-cpu
python-dotenv